+++ examples/inverse/plot_tf_dics.py
@@ -83,7 +83,8 @@ epochs_noise.apply_proj()
 epochs_noise = epochs_noise[:len(epochs.events)]
 
 # Read forward operator
-forward = mne.read_forward_solution(fname_fwd, surf_ori=True)
+forward = mne.read_forward_solution(fname_fwd)
+forward = mne.convert_forward_solution(forward, surf_ori=True)
 
 # Read label
 label = mne.read_label(fname_label)