+++ mne/inverse_sparse/_gamma_map.py
@@ -1,12 +1,11 @@
 # Authors: Alexandre Gramfort <alexandre.gramfort@telecom-paristech.fr>
 #          Martin Luessi <mluessi@nmr.mgh.harvard.edu>
 # License: Simplified BSD
-from copy import deepcopy
 
 import numpy as np
 from scipy import linalg
 
-from ..forward import is_fixed_orient, _to_fixed_ori
+from ..forward import is_fixed_orient, convert_forward_solution
 
 from ..minimum_norm.inverse import _check_reference
 from ..utils import logger, verbose, warn
@@ -242,8 +241,8 @@ def gamma_map(evoked, forward, noise_cov, alpha, loose=0.2, depth=0.8,
 
     # make forward solution in fixed orientation if necessary
     if loose is None and not is_fixed_orient(forward):
-        forward = deepcopy(forward)
-        _to_fixed_ori(forward)
+        forward = convert_forward_solution(
+            forward, surf_ori=True, force_fixed=True, copy=True, use_cps=True)
 
     if is_fixed_orient(forward) or not xyz_same_gamma:
         group_size = 1